+++ src/manifold/src/shared.h
@@ -192,7 +192,7 @@ Vec<TmpEdge> inline CreateTmpEdges(const Vec<Halfedge>& halfedge) {
   for_each_n(autoPolicy(edges.size()),
              zip(edges.begin(), halfedge.begin(), countAt(0)), edges.size(),
              Halfedge2Tmp());
-  int numEdge =
+  size_t numEdge =
       remove_if<decltype(edges.begin())>(
           autoPolicy(edges.size()), edges.begin(), edges.end(), TmpInvalid()) -
       edges.begin();
